oxp-sensors: hwmon: Add GPD Devices
Adds GPD Devices with EC Fan control. GPD devices don't have a separate enable register. The PWM register is used for
this purpose, with 0 as auto, 1-244 as a manual value, and+ 245-255 are undefined. We scale pwm1 0-255 and treat PWM
enable as setting pwm1 to manual at 70% as a common sense default.