ANDROID: fuse: Keep attributes consistent with Passthrough
If writeback cache is enabled, and we attempt to use fuse passthrough,
fuse will invalidate attributes on write, but will then ignore when the
server attempts to update those attributes. Since the kernel is the
arbiter of truth with writeback cache, passthrough holds the
responsibility to keep these values updated.

Signed-off-by: Daniel Rosenberg <drosen@google.com>
Bug: 333497409
Test: atest android.scopedstorage.cts.general.ScopedStorageDeviceTest
Change-Id: I1ebc0d06f3d2b41a9cfb2af35f27989d25d96032